+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to `ixudra/render` will be documented in this file
 
+## 1.0.0 2019-05-13
+### Added
+- Recursive translation functions in a given locale
+
+### Updated
+- Change default locale to null instead of empty string
+
 ## 0.1.3 - 2017-11-16
 ### Added
 - Added default formats to date/time functions